Second draft
Add more information about script mode and TensorFlow serving containers
